+++ NEWS.md
@@ -6,6 +6,10 @@
 * length_sort(): Sorts vector elements by character length.
 * length_order(): Sorts data.frames by character length of elements in specified
 column(s).
+* is_curie(): Tests for CURIEs in character vectors, according to a specified
+definition that _always_ conforms to W3C CURIE Syntax 1.0.
+
+### Updated
 * drop_blank(): Now a generic with `character` and `list` methods.